PRM-341 Criar barra de progresso para impressão de relatórios

+++ src/functions/utils.php
@@ -286,7 +286,7 @@ class utils
         return $html;
     }
 
-    public static function clear_temp_folder($ps_time = "1 minute")
+    public static function clear_temp_folder($ps_time = "1 minute", $part_filename = ""): void
     {
         $va_files = scandir(config::get(["pasta_media", "temp"]));
 
@@ -297,10 +297,29 @@ class utils
                 continue;
             }
 
+            if ($part_filename != "" && strpos($vs_file, $part_filename) === false)
+            {
+                continue;
+            }
+
             if (filemtime(config::get(["pasta_media", "temp"]) . $vs_file) < strtotime($ps_time))
             {
                 unlink(config::get(["pasta_media", "temp"]) . $vs_file);
             }
+        }
+    }
+
+    public static function callback_progress($vs_file_name, $vn_progress)
+    {
+        $vs_temp = config::get(["pasta_media", "temp"]);
+
+        $vs_file_path = $vs_temp . $vs_file_name . ".progress";
+
+        file_put_contents($vs_file_path, $vn_progress);
+
+        if (file_exists($vs_temp . $vs_file_name . ".stop")) {
+            utils::clear_temp_folder("1 minute", $vs_file_name);
+            exit();
         }
     }
